+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Changes in version 4.?.? - 2020-0?-??
+
+
 # Changes in version 4.3.0 - 2020-01-15
 
  * Medium changes